requirements.txt
# requirements.txt
numpy
pandas
keras-tcn
matplotlib
seaborn
scikit-learn
tensorflow>=2.9.0
argparse
streamlit